:building_construction: Restructure util folder
