+++ src/util/format.rs
@@ -1,10 +1,11 @@
-use super::convert::{days_to_date, days_to_doy, days_to_wyear, nanos_to_time};
-use crate::{
-    shared::{
-        MONTH_ABBREVIATED, MONTH_NARROW, MONTH_WIDE, NANOS_PER_SEC, SECS_PER_DAY, SECS_PER_HOUR,
-        SECS_PER_MINUTE, WDAY_ABBREVIATED, WDAY_NARROW, WDAY_SHORT, WDAY_WIDE,
-    },
-    util::convert::days_to_wday,
+use crate::shared::{
+    MONTH_ABBREVIATED, MONTH_NARROW, MONTH_WIDE, NANOS_PER_SEC, SECS_PER_DAY, SECS_PER_HOUR,
+    SECS_PER_MINUTE, WDAY_ABBREVIATED, WDAY_NARROW, WDAY_SHORT, WDAY_WIDE,
+};
+
+use super::{
+    date::convert::{days_to_date, days_to_doy, days_to_wday, days_to_wyear},
+    time::convert::nanos_to_time,
 };
 
 /// Formats string parts based on https://www.unicode.org/reports/tr35/tr35-dates.html#table-date-field-symbol-table
