+++ src/util/mod.rs
@@ -1,5 +1,6 @@
-pub(crate) mod convert;
+pub(crate) mod date;
 pub(crate) mod format;
 pub(crate) mod leap;
-pub(crate) mod manipulation;
+pub(crate) mod offset;
 pub(crate) mod parse;
+pub(crate) mod time;
